tests/wpt/tests/infrastructure/testharness/full.stop/full-stop.html

<!DOCTYPE HTML>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(function () {
      assert_equals(this.name, 'full-stop', 'Check that test name does not contain part of the path (.stop)');
  });
</script>
